Cycle Detection in an Undirected Graph
The algorithm steps are as follows:

For BFS traversal, we need a queue data structure and a visited array. 

Push the pair of the source node and its parent data (<source, parent>) in the queue, and mark the node as visited. The parent will be needed so that we don’t do a backward traversal in the graph, we just move frontwards. 

Start the BFS traversal, pop out an element from the queue every time and travel to all its unvisited neighbors using an adjacency list.

Repeat the steps either until the queue becomes empty, or a node appears to be already visited which is not the parent, even though we traveled in different directions during the traversal, indicating there is a cycle.

If the queue becomes empty and no such node is found then there is no cycle in the graph.